checker [ load-balancer, checking skeleton code], derecho [refactored leader-selection, null message, get_min_idx_rec]
load-balancer : Added code for load-balancing, which distributes load/requests evenly on the nodes in the system, and thus reduces the number of null messages sent in the system.
Skeleton-code: Added code to write statistics of the number of messages of each type sent in the system, and placeholders for checking. <To-Do> check the number of message count.
Derecho - leader-selection: Refactored deadlock leader-selection code, and moved the call to the run method
Null-message - Resolved issue of holes in the slots when null message was not sent due to full buffers
get_min_idx_rec - changed the lagging node to min from max, added comments of failure.