added back active field to resolve issue while failures
When failures happened, some of the nodes did not receive the view_change message from all the nodes, while one of the nodes started processing requests. This resulted in the system getting stuck because this message was dropped on nodes still waiting to start their epoch. This code adds a synchronization condition to overcome the issue.